+++ js/utils.js
@@ -33,7 +33,17 @@ const fillString = (string, minLength, additionalSymbols) => {
   return string;
 };
 
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
 checkLength('test', 1);
 isPalindrome('test');
 getInteger('test');
 fillString('test', 1, 'test');
+
+export {getRandomInteger};
